feat: decouple example downloading
