+++ ~Samples/AvatarLod/Avatar Configuration LOD0.asset
