+++ ~Samples/AvatarLod/Avatar Configuration LOD1.asset
