+++ ~Samples/AvatarLod/Avatar Configuration LOD2.asset
