+++ ~Samples/AvatarLod/AvatarLodExample.unity.meta
