+++ ~Samples/AvatarRender/AvatarRenderExample.cs
