+++ ~Samples/AvatarRender/AvatarRenderExample.cs.meta
