+++ ~Samples/MultipleAvatarLoadingExample.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 69dcfd12ccd44feb865f4afe67dc5f41
+timeCreated: 1669378324
