+++ ~Samples/MultipleQualityAvatarLoading/MultipleQualityAvatarLoadingExample.unity
