+++ ~Samples/MultipleQualityAvatarLoading/MultipleQualityAvatarLoadingExample.unity.meta
