+++ ~Samples/MultipleQualityAvatarLoading/QualityContainer.prefab
