+++ ~Samples/MultipleQualityAvatarLoading/QualityContainer.prefab.meta
